add seperate component for adding players before a game
