+++ src/pages/Crime.js
@@ -5,7 +5,7 @@ import DeleteForever from '@mui/icons-material/DeleteForever';
 import Button from '@mui/material/Button';
 import AddCircleOutline from '@mui/icons-material/AddCircleOutline';
 import DrinkingGame from '../classes/DrinkingGame';
-
+import RegisterPlayerField from '../components/RegisterPlayerField';
 const Crime = () => {
   const [players, setPlayers] = useState(JSON.parse(localStorage.getItem('players')) || ["Halvor","Martin","Edvard","Adrian"]);
   const [gameStarted, setGameStarted] = useState(JSON.parse(localStorage.getItem('gameStarted')) || false);
@@ -41,17 +41,6 @@ const Crime = () => {
     localStorage.removeItem('game');
   };
 
-  function deletePlayer(index) {
-    const newPlayers = [...players];
-    newPlayers.splice(index, 1);
-    setPlayers(newPlayers);
-  }
-
-  function addPlayer() {
-    const newPlayers = [...players, ""];
-    setPlayers(newPlayers);
-  }
-
   function startGame() {
     const filteredPlayers = players.filter(player => player !== "");
     if (filteredPlayers.length < 3) {
@@ -65,11 +54,6 @@ const Crime = () => {
     setGameState(intialQuestion);
   }
 
-  function handlePlayerChange(index, value) {
-    const newPlayers = [...players];
-    newPlayers[index] = value;
-    setPlayers(newPlayers);
-  }
 
   function newQuestion() {
     if (game) {
@@ -84,14 +68,7 @@ const Crime = () => {
       setGameState(previous);
     }
   }
-  /*
-  useEffect(() => {
-    if (game) {
-      newQuestion();
-    }
-  }, [game]);
 
-  */
 
   return (
     <div className="App-body">
@@ -110,30 +87,12 @@ const Crime = () => {
       </>
     ) : (
       <>
-        <p>Velkommem til Crime!</p>
-        <div className="inputFieldContainer">
-          {players.map((player, index) => (
-            <div key={index}>
-              <TextField
-                id={`outlined-basic-${index}`}
-                variant="outlined"
-                className="inputField"
-                value={player}
-                onChange={(e) => handlePlayerChange(index, e.target.value)}
-              />
-              <IconButton size="large" className="deleteButton" onClick={() => deletePlayer(index)}>
-                <DeleteForever />
-              </IconButton>
-            </div>
-          ))}
-          <IconButton size="large" className="addButton" onClick={addPlayer} >
-            <AddCircleOutline />
-          </IconButton>
-        </div>
-        <Button variant="contained" size="large" onClick={startGame}>Lets do crime!</Button>
-        </>
+      <h1>Crime</h1>
+      <RegisterPlayerField players={players} setPlayers={setPlayers} startGame={startGame} />
+      </>
     )}
     </div>
+    
   );
 };
