+++ src/pages/SpinWheel.js
@@ -6,7 +6,7 @@ import { Button, TextField, IconButton } from '@mui/material';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import { useNavigate } from 'react-router-dom';
-
+import RegisterPlayerField from '../components/RegisterPlayerField.js';
 const SpinWheel = () => {
     const [players, setPlayers] = useState(JSON.parse(localStorage.getItem('players')) || []);
     const [game, setGame] = useState(null);
@@ -76,24 +76,7 @@ const SpinWheel = () => {
     if (!gameStarted || !game) return (
         <div className="App-body">
             <h1>Spin The Wheel Game</h1>
-            {players.map((player, index) => (
-                <div key={index} style={{ display: 'flex', alignItems: 'center', marginBottom: '10px' }}>
-                    <TextField
-                        variant="outlined"
-                        value={player}
-                        onChange={(event) => handlePlayerChange(index, event)}
-                        style={{ marginRight: '10px' }}
-                    />
-                    <IconButton onClick={() => deletePlayer(index)}>
-                        <DeleteForeverIcon />
-                    </IconButton>
-                </div>
-            ))}
-            <IconButton onClick={addPlayer} style={{ marginTop: '10px' }}>
-                <AddCircleOutlineIcon />
-            </IconButton>
-            <br />
-            <Button variant="contained" onClick={startGame} style={{ marginTop: '20px' }}>Start Game</Button>
+            <RegisterPlayerField players={players} setPlayers={setPlayers} startGame={startGame} />
         </div>
     );
